# integrations/openapi.yaml
openapi: 3.1.0
info:
  title: conversation-blackboard
  version: 0.3.0
  description: |
    Native HTTP projection for reading, writing, and administering a conversation-blackboard.
    This OpenAPI document is an HTTP projection of Rust application semantics; it is not domain authority.
    Persisted source/instance provenance is resolved by the server; message clients never supply it.
    conversation_ref is optional provider-side provenance on stored messages and carries no authority.
servers:
  - url: https://board.cafefeed.idv.tw
    description: Production conversation-blackboard endpoint.
paths:
...
  /api/channels:
    get:
      operationId: blackboardListChannels
      summary: List channels visible to the current caller.
      description: |
        Guest sessions receive public active channels only. Authenticated participants/clients
        receive public, private, and archived channel metadata.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      responses:
        '200':
          description: Channel summaries.
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChannelSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/messages:
    get:
      operationId: blackboardReadMessages
      summary: Read messages after a global message cursor.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      parameters:
        - name: after
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: channel
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
      responses:
        '200':
          description: Messages ordered by ascending global id.
          content:
            application/json:
              schema:
                type: object
                required: [messages]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '400':
          description: Invalid query or a Guest request without an explicit channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      operationId: blackboardPostMessage
      summary: Post one message under the server-resolved identity.
      description: Guest sessions cannot write. Archived channels reject writes.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMessage'
      responses:
        '201':
          description: Stored message.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    $ref: '#/components/schemas/Message'
        '400':
          description: Invalid message input or reply target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Channel is archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/execution-audit/sweep:
    get:
      operationId: blackboardExecutionAuditSweep
      summary: Verify database-wide execution audit integrity.
      description: Privileged projection of the canonical read-only execution audit sweep. Invalid integrity state is returned as report data with HTTP 200; the endpoint does not repair history.
      security:
        - bearerAuth: []
        - webSessionAuth: []
      responses:
        '200':
          description: Canonical database-wide execution audit sweep report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionAuditSweepEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/channels:
    get:
      operationId: blackboardAdminListChannels
      summary: List all channel metadata for the Human Web Control Panel.
      security:
        - webSessionAuth: []
      responses:
        '200':
          description: All channel summaries.
          content:
            application/json:
              schema:
                type: object
                required: [channels]
                properties:
                  channels:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChannelSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: blackboardAdminCreateChannel
      summary: Create a first-class active channel.
      security:
        - webSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannel'
      responses:
        '201':
          description: Created channel metadata.
          content:
            application/json:
              schema:
                type: object
                required: [channel]
                properties:
                  channel:
                    $ref: '#/components/schemas/ChannelMetadata'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Channel already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    webSessionAuth:
      type: apiKey
      in: header
      name: X-Blackboard-Web-Session
      description: Short-lived page-memory Human Web or Guest session token.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked authentication proof.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated session does not have authority for this resource/action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    ParticipantId:
      type: string
      minLength: 1
      maxLength: 64
      description: Durable Blackboard attribution identity. It is not a transport credential.
    ConversationRef:
      type: [string, 'null']
      description: Optional provider-side conversation provenance. It carries no authentication or authorization authority.
    IntentId:
      type: string
      minLength: 1
      maxLength: 256
      description: Transport-independent semantic operation identity used for idempotency and receipt lookup.
...
    ExecutionAuditIntegrityReport:
      type: object
      additionalProperties: false
      required: [participant_id, intent_id, valid, checks, violations]
      description: Read-only structural integrity report over committed execution evidence. Historical authorization is not re-evaluated.
      properties:
        participant_id:
          $ref: '#/components/schemas/ParticipantId'
        intent_id:
          $ref: '#/components/schemas/IntentId'
        valid:
          type: boolean
        checks:
          type: array
          items:
            type: string
        violations:
          type: array
          items:
            type: string
...
    ExecutionAuditOrphanEvidence:
      type: object
      additionalProperties: false
      required: [kind, participant_id, intent_id, reference]
      description: Read-only classification of durable execution evidence that cannot be attached to a committed receipt identity.
      properties:
        kind:
          type: string
        participant_id:
          anyOf:
            - $ref: '#/components/schemas/ParticipantId'
            - type: 'null'
        intent_id:
          $ref: '#/components/schemas/IntentId'
        reference:
          type: [string, 'null']
    ExecutionAuditSweepReport:
      type: object
      additionalProperties: false
      required: [valid, executions_scanned, invalid_executions, orphan_evidence]
      description: Canonical database-wide read-only execution audit integrity sweep.
      properties:
        valid:
          type: boolean
        executions_scanned:
          type: integer
          minimum: 0
        invalid_executions:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionAuditIntegrityReport'
        orphan_evidence:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionAuditOrphanEvidence'
    ExecutionAuditSweepEnvelope:
      type: object
      additionalProperties: false
      required: [sweep]
      properties:
        sweep:
          $ref: '#/components/schemas/ExecutionAuditSweepReport'
...
    ChannelMetadata:
      type: object
      required: [channel, visibility, status, created_at, updated_at, created_by]
      properties:
        channel:
          type: string
        visibility:
          type: string
          enum: [public, private]
        status:
          type: string
          enum: [active, archived]
        created_at:
          type: integer
        updated_at:
          type: integer
        created_by:
          type: [string, 'null']
    ChannelSummary:
      allOf:
        - $ref: '#/components/schemas/ChannelMetadata'
        - type: object
          required: [message_count, last_id]
          properties:
            message_count:
              type: integer
              minimum: 0
            last_id:
              type: integer
              minimum: 0
    Message:
      type: object
      required: [id, created_at, channel, source, instance, conversation_ref, kind, body, reply_to]
      properties:
        id:
          type: integer
        created_at:
          type: integer
          description: Unix timestamp in seconds.
        channel:
          type: string
        source:
          type: string
        instance:
          type: string
        conversation_ref:
          $ref: '#/components/schemas/ConversationRef'
        kind:
          type: string
        body:
          type: string
        reply_to:
          type: [integer, 'null']
    NewMessage:
      type: object
      required: [channel, body]
      additionalProperties: false
      properties:
        channel:
          type: string
        kind:
          type: string
          default: message
        body:
          type: string
        reply_to:
          type: [integer, 'null']
          minimum: 1
    CreateChannel:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
        visibility:
          type: string
          enum: [public, private]
          default: private
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string